Move form to the evals app
